#1695: - merged RepositoryUtil.java into RepositoryType - updated tests
